refactor(test): split epub preview tests by package toc navigation and assets
